Refactor the location of all custom assertions
