+++ .eslintrc.js
@@ -25,7 +25,7 @@ module.exports = {
     // "consistent-return": "off",
     // "dot-notation": "off",
     "func-names": "off",
-    // "global-require": "off",
+    "global-require": "off",
     // "guard-for-in": "off",
     // "import/newline-after-import": "off",
     "import/no-dynamic-require": "off",
